refactor: memberType 변수명 변경으로 인한 수정
